refactor: JPA 모듈 추가
