+++ module-infra/persistence-db/src/main/kotlin/demo/kotlinpractice/member/repository/MemberJPARepository.kt
@@ -1,4 +1,4 @@
-package demo.kotlinpractice.domain.member.infra.repository
+package demo.kotlinpractice.member.repository
 
 import demo.kotlinpractice.domain.member.domain.Member
 import demo.kotlinpractice.domain.member.infra.MemberEntity
